feat: ✨ add `verify_is_dir()` (#578)
## Description

These changes are done to start the work of creating the
`create_resource_structure()`.

Closes #562 

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review. Focus on:

1. Do we want to use `pathlib` and not `os`?  --> Yes.
2. I have added a new file `verify_is_directory.py` since I wasn't sure
how we want to structure helper functions like this one. Where should it
be located? --> This is fine for now.
3. Tests: 
1. How do we want to name them? --> We don't need the name of the
function in the name of the test as long as we keep one test file per
function. Then the function name will be in the name of the test file.
    2. Am I missing a test case?

## Checklist

- [X] Added or updated tests
- [X] Tests passed locally
- [X] Linted and formatted code
- [X] Updated documentation (yes: docstrings)

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>