docs: :building_construction: function diagram for `create_properties_template()` (#600)
## Description

These changes add the diagram for creating the properties template of
the official Data Package Spec. I think it's possible to download the
full spec from one URL. But maybe this will need two functions, one for
the package and one for the resources. Thoughts? Alternatively, I could
add a URL or options arg to select which spec to download, the package
or the resource.

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>