docs: :memo: add description of `extract_resource_properties()` (#603)
## Description

These changes add a description for an extraction of properties function
for data files. This function needs to be exposed to the user,
especially since it needs to be exposed to be used properly for the web
app. It also simplifies the `create_resource_properties()` function a
lot!

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>
Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>