feat: ✨ add `create_resource_structure()` (#610)
## Description

Bringing together the smaller functions to make an external facing
function.

Closes #568 

Relates to #533 (bc I have added a `__init__.py` in `core` with
`create_resource_structure()`)

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review. 

Focus on: 

1. How do we want to import internal functions to user-facing functions
like `create_resource_structure()`? This also relates to general
refactoring; right now all core functions are in the same folder which
will become messy as we add new functions. Thoughts at this point? Do we
want to keep as is for now and then wait for a refactoring session later
on?

## Checklist

- [X] Added or updated tests
- [X] Tests passed locally
- [X] Linted and formatted code
- [X] Updated documentation (=docstrings)

---------

Co-authored-by: Kris Beicher <112945740+K-Beicher@users.noreply.github.com>
Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>