chore: :children_crossing: limit functions accessible when importing all (#626)
## Description

These changes sets up the structure so that whenever `from sprout.core
import *` is used, it only gives access to the functions we want. So
far, I've only added commented code, but we can use this also as a nice
visual to see our progress :tada:

Closes #533

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.