feat: :sparkles: auto-generate property dataclasses based on spec (#692)
## Description

This now generates dataclasses based on the Data Package spec using a
template that produces better docstrings and appeases the linter for the
most part. Not counting the fields and anything under those, I got it
down to about 10 classes.

I included both the generated file and a `properties.py` file with minor
edits (adding a resource `id`, some formatting fixes, some missing
docstrings).

What remains to be done:

- Some docstrings are missing in the spec, these will need to be filled
in by hand. This could include a proofreading of the existing
docstrings.
- We could add the default values from the other
[PR](https://github.com/seedcase-project/seedcase-sprout/pull/676).
- We need to write a reasonable class for table schema fields and add
this in.
- Have a think about whether we want to add any tests around the
generation logic. Seeing that we need to proofread the classes every
time we generate them anyway and that we probably won't be regenerating
them all that often, I think testing the generation logic properly would
be kinda misplaced effort. Maybe running it on a bare-bones (= fake)
JSON schema and checking that it generates what we want / runs at all
could work?

~~This is to show what it would look like if we used
[datamodel-code-generator](https://github.com/koxudaxi/datamodel-code-generator)
to extract class definitions from the
[spec](https://datapackage.org/profiles/2.0/datapackage.json). I'm sure
the output can be customised a bit more and where it puts the file would
need to be decided etc., but I wanted to get some feedback on this rough
draft to see if this is something we should work more on.~~

~~We can also have a script post-process the output if we think it's
useful, but there are some things we want to rearrange.~~


## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

Focus on the output generated by the library -- would this make our life
easier?

## Checklist

- [?] Added or updated tests
- [x] Tests passed locally
- [x] Linted and formatted code
- [x] Build passed locally
- [x] Updated documentation