feat: :sparkles: add `FieldProperties` and `ConstraintsProperties` (#717)
## Description

This adds dataclasses for `FieldProperties` and `ConstraintsProperties`.

The direct translation of the spec into classes would map each field
type (for each data type) to a separate class. These classes are similar
but some fields have more properties (e.g. `categories` is available
only for integer and string fields) and some properties are structured
differently for different fields (e.g. different constraints are
applicable to string fields and integer fields).
[Here](https://datapackage.org/standard/table-schema/#field)'s a nice
description of the properties.

In `FieldProperties` I included all properties (except `rdfType`),
accepting that this way the onus of not constructing meaningless fields
(e.g. a string field with a maximum value of the integer 12) is on the
user / downstream consumer. I constructed `ConstraintsProperties`
similarly.

The docstrings were copied from the spec.

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.


## Checklist

- [x] Added or updated tests
- [x] Tests passed locally
- [x] Linted and formatted code
- [x] Build passed locally
- [x] Updated documentation

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>