docs: :memo: improve on the docstrings in `*Properties` classes (#718)
## Description

This improves the docstrings in the properties classes. Some missing
descriptions are added and some errors are corrected. I've also read
over the [website](https://datapackage.org/standard/data-package/) that
contains more descriptive documentation than the JSON schema and moved
our docs to that version if I thought it was nicer (pretty arbitrarily).

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs a medium-depth review.

## Checklist

- [x] Added or updated tests
- [x] Tests passed locally
- [x] Linted and formatted code
- [x] Build passed locally
- [x] Updated documentation

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst@gmail.com>