feat: :sparkles: add `read_json()` (#746)
## Description

This adds a simple `read_json()` function, which loads the contents of a
JSON file into a dictionary / list.
This is basically the `read_properties` function without any
properties-specific validation. This way, it's not tied to the
`datapackage.json` file in any way, so I used a more general name. If we
do want to add properties-specific validation, we can still build on
this function.

Not sure if closes but related to #711 

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.


## Checklist

- [x] Added or updated tests
- [x] Tests passed locally
- [x] Linted and formatted code
- [x] Build passed locally
- [x] Updated documentation