feat: ✨ add `path_*()` functions (with their `verify_*()`'ers) (#730)
## Description
  
I decided to have a script for path functions related to packages and
one related to resources. I found that it became too long - for my taste
at least - with the docstrings otherwise.
  
To add existing ID's to the raised errors, I have added two verify
functions.
  
To test the path functions, I have created some helper functions to
create the directory structure of packages and resources.

Closes #714  

## Reviewer Focus
  
<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.  

## Checklist

- [X] Added or updated tests
- [X] Tests passed locally
- [X] Linted and formatted code
- [X] Build passed locally
- [X] Updated documentation (= docstrings)

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst@gmail.com>