docs: :building_construction: internal diagram for `write_resource_properties()` (#751)
## Description

The internal flow is slightly different from the
`write_package_properties()` in that the `datapackage.json` already has
to exist because we need to join the existing properties to correctly
add the resource to it.

I'm not certain if we need a check or function to put the resource in
the right place in the spec. But let's see how the function gets
developed and used.

Closes #647

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.