fix: :bug: sort ids before returning them (#767)
## Description

This sorts IDs before returning them. It seems to be the case that the
order in which dirs are returned is not determined by Python, which can
lead to some inconsistent behaviour. This is causing some tests to fail
remotely, but not locally (for me at least).

Even more interestingly, the tests are run for PRs where neither one of
the participating branches even has the commit with the test (🤔). I have
no idea why and this does nothing to resolve that.

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

Focus on CHANGES.

## Checklist

- [x] Added or updated tests
- [x] Tests passed locally
- [x] Linted and formatted code
- [x] Build passed locally
- [x] Updated documentation

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>