docs: :sparkles: add content to landing page (#772)
## Description

This is my first stab at creating the landing page. I have tried to keep
it as short as possible (i.e., not a lot of text), so that people wont
get overwhelmed when they enter the landing page.

I think that landing pages shouldn't contain that much content on its
own, but rather point to the relevant resources (like how to install and
to the tutorials).

What do you think? :)))) 

Note: I will add some changes to the theme to make e.g., the buttons (in
my opinion) a bit more modern-looking.

A screenshot from my local version:


![image](https://github.com/user-attachments/assets/011f72b5-764a-4aaa-a9ae-de283e40d689)

Closes #725 

## Reviewer Focus

<!-- Select quick/in-depth as necessary -->
This PR needs a in-depth review.

## Checklist

- [X] Formatted text
- [X] Build passed locally
- [X] Updated documentation