refactor: 🔥 remove everything Django-related (#848)
## Description

This PR removes all folders, files, tests, dependencies, just recipes,
etc. that are related to our use of Django. We are not currently using
Django and if we decide to use it again later on, the work removed here
would need a complete refactor/rebuild anyway, since the Django app was
created before we decided to rely on the Data Package standard.

Closes #845 

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review. Is there something I have forgotten to
remove? 🔥

## Checklist

- [X] Added or updated tests
- [ ] Tests passed locally (no - see #851 for fix of the failing tests)
- [X] Linted and formatted code

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>