chore: :fire: no longer need these files (#877)
## Description

This PR removes most of the PlantUML files and docs that focused on the
web app. They were super useful at the time and help evolve our
understanding of what we wanted to do and build, but are now very
out-dated with our current state. So, like the Django removal, these
remove all these docs related to the Django and web app stuff.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.