docs: :memo: describe split between `lib` and `core` (#872)
## Description

As I was writing the guide on creating data resources and late night
swirling thoughts, combined with potential users who would code in
Python, I thought that there should actually be Python code that mimicks
what the CLI will do... meaning code that has a lot more assumptions and
opinions tied to it. Based on discussions we've had over the months, I
see a lot of value in having a highly flexible `core`, that would allow
us to easily integrate e.g. plugins/extensions, not be tied to the
specific folder and file structure we prescribe, and allow others to
customize things a lot more to their needs. But there is also a lot of
value in having some code that is very opinionated and simple. For
instance, not needing to pass around `path_*()` for all functions and
just assume "properties will always be the root of the package", etc.

Closes #850 and closes #869

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>