feat: :sparkles: add simple helper functions (#962)
## Description

This PR adds some simple helper functions that will be used when running
more complex checks in Sprout.

As a rule, I tried to write simple checks in such a way that they only
check a single thing. So if a function checks whether a field is blank,
it won't fail if the field is missing or is of the wrong type, because
there are separate checks for that. That said, it's always possible to
argue that this doesn't make sense for some functions -- so feel free to
argue if you think any checks should behave differently!

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>