docs: :fire: don't actually need the view action, have the README (#1012)
## Description

I mainly wrote the view and list functions in the context of the CLI or
app, where it was more interactive and it was multi-user. But since
we're focusing on core and lib, and since we have the information in
human-readable form in the README, it doesn't make sense to have these
`view_` and `list_` interfaces. When we do develop CLI, we can implement
in cli if necessary, but I think the functionality with making the
README will be sufficient for most purposes.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Updated documentation