refactor: :recycle: commonise Sprout checks (#998)
## Description

This PR shares logic across Sprout checks.
As all top-level Sprout checks call a check function from the `checks`
package and then add Sprout-specific logic, there was lots of overlap
between them. Here, I left that logic in `check_properties`, and made
`check_resource_properties` and `check_package_properties` rely on it.
The underlying idea being "do all checks and ignore irrelevant errors".
This works with the existing error-ignoring strategy.

In our codebase, as it is currently, this made more sense than to do it
the other way and build `check_properties` by calling
`check_resource_properties` and `check_package_properties`. But feel
free to argue otherwise!

Doing it this way, we're only using one check function from the `checks`
package (the simplest one), so that's nice.

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`