docs: :memo: move user types over into user personas document (#994)
## Description

Continuing the design docs clean up. I simplified the user roles and
then made some user personas with a non-exhaustive list of things they
may have as goals, needs, challenges, and wants. These can sometimes be
useful to communicate who the expected audience is for this package.

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>