docs: :memo: pseudo code and docstring for `write_resource_parquet()` (#816)
## Description

Based on @martonvago's suggestion, I'll write things in "pseudocode"
from now on. But instead of pseudocode, I will write an outline of the
Python function with how I think it might flow inside. Plus, I can write
the full docstrings inside, so you all don't need and we don't need to
move it over from the Quarto doc. **I have NOT ran this, tested it, or
did any execution, this is purely how I think it might work**, hence
"pseudo" :stuck_out_tongue:. I'll add some comments directly to the code
in the PR.

Closes #642

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Updated documentation

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>