docs: :building_construction: expand and refine naming scheme (#1070)
## Description

I've been thinking a bit more about the naming scheme. As we've
developed more and used it and added the guide, some things feel a bit
rough/has tension. Especially with how we've named things. Comments
about each change is added to the PR.

Closes #1059

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.