docs: :building_construction: revise to `build_resource_data()` to output a DataFrame (#1074)
## Description

After seeing @K-Beicher working with the data and making the guide, it
felt off to immediately save to Parquet. Given this is `core` and
developer-focused/highly flexible, I can imagine situations where there
may be some post-processing before generating to the Parquet file. So
I've modified this function to output to a DataFrame rather than save
directly to Parquet.

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>