refactor: :recycle: switch `create_package_structure()` to make properties only (#1047)
## Description

The `create_package_structure()` was no longer relevant, so I refactored
it to the `create_package_properties()`. This function takes properties
and a path as input and creates the `datapackage.json` file as well as
the folder for the data package.

I had to comment out several docstring examples since they depended on
the global structure rather than the "local-first" approach.

Closes #1033
Closes #958

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>
Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>