feat: :sparkles: add `path_readme()` (#1100)
## Description

This PR adds a `path_readme()` which was referred to in a TODO item in
#810
This function will probably have to be rewritten a bit when we fully
switch to the “local-first” approach, but I thought it made sense to add
it now, so we don’t forget about it.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [X] Added or updated tests
- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>