refactor: :recycle: major refactor of path functions to be local-first (#1114)
## Description

Converts all the path functions to be either local-first or multi-user.
In the process I consolidated some of the functions into either "local"
or "global" files.

I moved some of the smaller helper functions into the larger files, and
appended with `_`. I'm not a huge fan of how it looks, but given it is a
common Python practice and it helps visually separate external from
internal functions, I'm ok if we start doing that too.

Closes #1004

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>
Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>