refactor: :fire: remove `create_package_properties()` and replace it with `write_*()` (#1153)
## Description

Based on the flow diagram in PR #1151, the `create_package_properties()`
function isn't actually needed, so I've removed it and replaced all
cases with just `write_package_properties()`. This simplifies things a
lot! There wasn't needed a lot of change aside from renaming `create_`
to `write_` so this isn't a tough review.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>