docs: :building_construction: pseudocode for `_check_column_names()` (#1194)
## Description

I know this function has already been made, but it needs some tweaks
based on the general discussions around the check functionality. So I
decided to start here with the pseudocode of that to get on the same
page with this *before* we refactor the existing function.

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.