refactor: :recycle: clean split of `check_datapackage` sub-package (#1239)
## Description

This PR makes it so there is a cleaner split with the sub-package and
all other code. For instance, keeping a clear external function listing,
importing from the sub-package directly rather than from the
sub-modules, etc.

Closes #1224

<!-- Select quick/in-depth as necessary -->
This PR needs a bit of an in-depth review. It's mostly renaming and
moving files around.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`