test: :fire: remove duplicate tests on `check_*` functionality (#1256)
## Description

If we extensively test a functionality that another function uses
internally, we don't need to also duplicate the tests again in the other
function. Since the move to refactor `check_properties()` and having the
more extensive tests there, these tests are not needed any more as they
are duplicates.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Ran `just run-all`