refactor: :recycle: create a simpler map functional (#1270)
## Description

It bothers me to no end that Pythons built-in functionals are so
user-unfriendly. So this is the start of some refactoring I will do,
with simpler user-friendly versions of map that are inspired from the R
package purrr, which implements and designs functionals so nicely. So
these functionals are mostly a copy of what purrr has.

I will be adding other functionals here over time.

Closes #1267 

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: martonvago <57952344+martonvago@users.noreply.github.com>