feat: :sparkles: add example package (#1265)
## Description

This PR adds an example package class that can be used like:
```python
with ExamplePackage() as package_path:
    properties = read_properties(package_path.properties())
    ...
```

- can take a `with_resources` flag
- creates datapackage.json, README, resources folder, folder for one
resource (if `with_resources` is True)
- puts everything into a temporary folder
- cleans up after itself

It's easy to add more customisations. 
It's easy to add data batches and data.parquet if we want.

I'll update example code to use this after it's OK'd

Closes #1048 

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`