feat: :sparkles: finish up `PackagePath` and `resource_batch_files()` (#1282)
## Description

This PR finishes up `PackagePath`, especially `resource_batch_files()`.
I made it not throw an error if there are no batch files for the
resource because the other methods don't throw either. I think returning
`[]` when there are no batch files is in keeping with the spirit of the
class. But we can always change it!

I also did some general tidying.

This is groundwork for #1210

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`