build: :hammer: add coverage infrastructure and badge (#1289)
## Description

Adds coverage infrastructure, plus adds some badges to the README.

In other PRs, I updated the justfile to run the coverage with pytest. I
also moved out the arguments to pytest into the justfile so that it is
easier to sync with other (future) repos.

Closes #521, closes #847, closes #805

<!-- Select quick/in-depth as necessary -->
Doesn't really need a review.