refactor: :recycle: have a default method only for `PackageProperties` (#1304)
## Description

This PR removes the default method from all properties classes except
`PackageProperties`.

In `PackageProperties`, I changed the default method to be in line with
how it's used in the guide. Based on an old
[discussion](https://github.com/seedcase-project/seedcase-sprout/pull/1153#discussion_r2003604421).
Another approach would be to populate these fields automatically when
creating the object in the usual way, but I'm not sure if we want that.

Closes #1209

<!-- Select quick/in-depth as necessary -->
This PR needs a medium-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>