fix: :bug: resolve temp paths so tests pass on MacOS as well (#1322)
## Description

Two tests were failing for me locally and it seems to be bc of how MacOS
uses `/var` as a symlink to `/private/var` (thanks @martonvago 🌟 ).
So, these two tests were failing bc the paths are compared directly and
even though they refer to the same location, the paths aren’t identical.
I solved this by resolving the paths with `.resolve()`.

A different solution could be to change how we assert the paths. Like
`assert Path.cwd().samefile(package_path.path)`.

Thoughts?

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [X] Added or updated tests
- [X] Ran `just run-all`