build: :hammer: build PDF locally (#1325)
## Description

This PR adds a Quarto profile for rendering the website as a PDF.

I'm sure this can be improved a lot, but I don't know Quarto that well,
so I'm putting this up for review now instead of sitting on it longer.

I've also included some tweaks to improve the output, but it's not
perfect yet.
Outstanding issues:
- The `glossary` extension doesn't support PDF. Replacing the shortcode
in the text as a post-processing step (e.g. `{{< glossary "data
package">}}` >> `data package`, so that at least it reads normally) is
not as trivial as it seems. It's still possible, but I didn't want to
spend more time on it if we're moving away from this extension anyway.
- Very long code output goes off the page. I mitigated this somewhat by
auto-wrapping code output, but these two cases remain:
  1. super long words (like an absolute path) are not wrapped
2. `pprint`ed objects have lots of whitespace that pushes them off the
page regardless of wrapping.
Displaying big, nested properties nicely is a problem for the HTML
website as well, so maybe we need a common solution?
- We could have a nice book cover / first page. There are Quarto
extensions that claim to help with that, but I haven't tried them.

This is what it looks like now:

[seedcase-sprout.pdf](https://github.com/user-attachments/files/20184031/seedcase-sprout.pdf)



Closes #1314 

<!-- Select quick/in-depth as necessary -->
This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>