feat: :sparkles: add custom package name to `ExamplePackage` (#1309)
## Description

This PR adds a `package_name` parameter in `ExamplePackage` for
specifying a custom package name.
The package name is set as the name of the root folder of the package
and as the name of the package properties.

I'll use `ExamplePackage` in the resource guide in another PR.

<!-- Select quick/in-depth as necessary -->
This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>