refactor: :recycle: simplify and make `check_datapackage` self-contained (#1342)
# Description

This cleans up the `check_datapackage` sub-package a bit. Mainly by
moving all the internal functions into `internals.py`. I also moved the
`read_json()` into here to keep it self-contained and not reference
things outside of it. I also removed the tests on the internal
functions.

This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>