feat: :sparkles: add properties template (#1345)
# Description

This PR adds a template for package properties (and supporting code).

I imagine the flow for using this like:

1. In a new package without a `datapackage.json`: either copy the
template to `scripts/properties.py` or call
`create_properties_template()` from the package root.
2. Open the template and fill in the required fields. Required fields
are uncommented. Comment out and fill in any optional fields you need.
3. Run `main.py`. This updates `datapackage.json` .

Closes #1343

This PR needs an in-depth review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`