build: :heavy_plus_sign: add jinja2 dependency explicitly (#1372)
# Description

This PR adds jinja2 as a dependency.
When installing `example-seed-beetle` in non-dev-mode, I saw that jinja2
wasn't being installed because it is not declared as a dependency in
Sprout.

Maybe we haven't seen this before because jinja2 is a dependency of many
dev-only dependencies in Sprout and of commitizen in
`example-seed-beetle`. And `uv sync` installs dev dependencies.

Let me know if I'm wrong, but this is my understanding :sweat_smile: 


This PR needs a quick review.

## Checklist

- [x] Added or updated tests
- [x] Updated documentation
- [x] Ran `just run-all`