fix: :bug: change `fields_match` to be an array in `extract_resource_properties()` (#1392)
# Description

It's supposed to be a string, but there's an error in the JSON schema.
According to a PR in the [data package
repo](https://github.com/frictionlessdata/datapackage/pull/1051), this
should be fixed in datapackage V2.1 🤞

I encountered this issue bc the website build fails bc there was an
error in `resources.qmd` when writing the properties.

This PR needs a quick review.

## Checklist

- [X] Ran `just run-all`